fix (S3): incorrect ModifiedAt timestamps for S3 (#200 - #204)
The problem addressed here is that the FileInfo struct's FTime field is
set in S3 backend. This was being passed as milliseconds epoch. That
value was being passed into `time.Unix(x, 0)` which accepts arguments as
either/both (seconds, nanoseconds).

By passing milliseconds to this function expecting seconds, we have
wildly incorrect modified at timestamps.

I tested this against the same bucket that was problematic before and
it's now showing correctly :).