fix (s3): fix remove a single object (#225)
Objects, as well as buckets are removed basing on objecs list received
from client. As the objects are fetched by Prefix, the request for
removing object 'foo' will remove all 'foo*' objects in this bucket.

For instance, having bucket with objects like so:

    awesomebucket/
    ├── foo
    ├── foobar
    └── thing

Rm("awesomebucket/foo") will have effect:

    awesomebucket/
    └── thing

This change fixes this bug by recognizing if single object has to be
removed or the entire bucket. For single object, we don't need to walk
through directories and can request to remove directly.