fix (#229): remove header mutation
On IOS, mutate an HTTP response would throw with the error "Headers object's guard is 'immutable". The idea of that mutation was to provide a way in frontend land to detect when a response is fresh or stale from a cache. As of today, we don't use that information so we removed the associated code