Allow to customize OnlyOffice callback URL
In some setups, the automatic detection of the callback URL where the
OnlyOffice container can reach filestash is wrong. It may be that IPv6
used instead or that the public IP address reaches the reverse proxy
which might not allow OnlyOffice to communicate. In controlled
environments, it's best to setup the callback URL manually as it is
known at configuration time.