fix (login): loader before redirection
When we have a single storage selected, we automatically redirect the
user to it as we don't need to wait for a user action. The thing is we
used to have this UI being shown with a button that didn't make sense.
Instead we want to show the loader when we reach those condition