fix (htpasswd): hash acceptance rules - #821
For htpasswd entries where hash is an actual valid hash (SHA etc.), the hash
itself should not be accepted as password. Otherwise, obtaining the htpasswd
file gives an attacker access to all accounts withoutneeding to crack/reverse
hashes.

This commit tries to ensure that hash==password is only accepted if hash is
not a valid hash (plaintext password fallback).